% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCollection.R
\name{createCollection}
\alias{createCollection}
\title{Creates a collection.}
\usage{
createCollection(connectionInfo, collectionId, throughput = "",
  partitionKey = "", indexingPolicy = "", databaseId = "",
  consistencyLevel = "", sessionToken = "", userAgent = "")
}
\arguments{
\item{connectionInfo}{A DocumentDB connection info object generated with getDocumentDBConnectionInfo().}

\item{collectionId}{The ID of the collection to create.}

\item{throughput}{Optional. The user specified throughput for the collection expressed in units of 100 request units per second. This can be between 400 and 250,000 (or higher by requesting a limit increase). If the value is over 10,000, then the collection must include a partitionKey definition. If the value is equal to or under 10,000, then the collection must not include a partitionKey definition.}

\item{partitionKey}{Optional. This value is used to configure the partition key to be used for partitioning data into multiple partitions. If the throughput value is over 10,000, then the collection must include a partitionKey definition. If the throughput value is equal to or under 10,000, then the collection must not include a partitionKey definition.}

\item{indexingPolicy}{Optional. This value is used to configure indexing policy. By default, the indexing is automatic for all document paths within the collection. To specify a custom indexing policy pass a string with a JSON element, eg. "indexingPolicy": {...}. See \url{https://docs.microsoft.com/en-us/rest/api/documentdb/create-a-collection} for more details.}

\item{databaseId}{Optional. The ID of the database in which to create the collection. If no database ID is specified, the ID will be taken from the connection info.}

\item{consistencyLevel}{Optional. The consistency level override. The valid values are: Strong, Bounded, Session, or Eventual (in order of strongest to weakest). The override must be the same or weaker than the account's configured consistency level.}

\item{sessionToken}{Optional. A string token used with session level consistency. For more information, see \href{https://azure.microsoft.com/en-us/documentation/articles/documentdb-consistency-levels}{Using consistency levels in DocumentDB}.}

\item{userAgent}{Optional. A string that specifies the client user agent performing the request. The recommended format is {user agent name}/{version}. For example, the official DocumentDB .NET SDK sets the User-Agent string to Microsoft.Document.Client/1.0.0.0. A custom user-agent could be something like ContosoMarketingApp/1.0.0.}
}
\value{
Some information extracted from the REST API response such as request charge and session token.
}
\description{
Creates a collection.
}
\examples{
# load the documentdbr package
library(documentdbr)

# setup connection infos
myDocDB <- getDocumentDBConnectionInfo(
  accountUrl = "https://somedocumentdbaccount.documents.azure.com",
  primaryOrSecondaryKey = "t0C36UstTJ4c6vdkFyImkaoB6L1yeQidadg6wasSwmaK2s8JxFbEXQ0e3AW9KE1xQqmOn0WtOi3lxloStmSeeg==",
  databaseId = "MyDatabase"
)

# create the MyCollection collection with default throughput (within MyDatabase as specified in the connection info)
# note: alternatively, the database id can be specified as parameter
createCollection(myDocDB, "MyNewDatabase")
}
