% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteCollection.R
\name{deleteCollection}
\alias{deleteCollection}
\title{Deletes a collection.}
\usage{
deleteCollection(connectionInfo, collectionId, databaseId = "",
  consistencyLevel = "", sessionToken = "", userAgent = "")
}
\arguments{
\item{connectionInfo}{A DocumentDB connection info object generated with getDocumentDBConnectionInfo().}

\item{collectionId}{The ID of the collection to delete.}

\item{databaseId}{Optional. The ID of the database in which to delete. If no database ID is specified, the ID will be taken from the connection info.}

\item{consistencyLevel}{Optional. The consistency level override. The valid values are: Strong, Bounded, Session, or Eventual (in order of strongest to weakest). The override must be the same or weaker than the account's configured consistency level.}

\item{sessionToken}{Optional. A string token used with session level consistency. For more information, see \href{https://azure.microsoft.com/en-us/documentation/articles/documentdb-consistency-levels}{Using consistency levels in DocumentDB}.}

\item{userAgent}{Optional. A string that specifies the client user agent performing the request. The recommended format is {user agent name}/{version}. For example, the official DocumentDB .NET SDK sets the User-Agent string to Microsoft.Document.Client/1.0.0.0. A custom user-agent could be something like ContosoMarketingApp/1.0.0.}
}
\value{
Some information extracted from the REST API response such as request charge and session token.
}
\description{
Deletes a collection.
}
\examples{
# load the documentdbr package
library(documentdbr)

# setup connection infos
myDocDB <- getDocumentDBConnectionInfo(
  accountUrl = "https://somedocumentdbaccount.documents.azure.com",
  primaryOrSecondaryKey = "t0C36UstTJ4c6vdkFyImkaoB6L1yeQidadg6wasSwmaK2s8JxFbEXQ0e3AW9KE1xQqmOn0WtOi3lxloStmSeeg==",
  databaseId = "MyDatabase"
)

# deletes the MyCollection collection (within MyDatabase as specified in the connection info)
# note: alternatively, the database id can be specified as parameter
deleteCollection(myDocDB, "MyCollection")
}
