% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteDocument.R
\name{deleteDocument}
\alias{deleteDocument}
\title{Deletes a document from a collection.}
\usage{
deleteDocument(accountUrl, primaryOrSecondaryKey, databaseId, collectionId,
  documentId, partitionKey = "")
}
\arguments{
\item{accountUrl}{The URI of the DocumentDB account.}

\item{primaryOrSecondaryKey}{The master key to authenticate.}

\item{databaseId}{The ID of the database to query.}

\item{collectionId}{The ID of the collection to query.}

\item{documentId}{The ID of the document to delete.}

\item{partitionKey}{Optional. The partition key value for the document to be deleted. Must be included if and only if the collection is created with a partitionKey definition.}
}
\value{
Some information extracted from the REST API response such as request charge and session token.
}
\description{
Deletes a document from a collection.
}
\examples{
# deletes the document with id fe7718ad-0000-4f42-cf5a-e2d79d2156df
library(documentdbr)
deleteResult <-
  deleteDocument(
    accountUrl = "https://somedocumentdbaccount.documents.azure.com",
    primaryOrSecondaryKey = "t0C36UstTJ4c6vdkFyImkaoB6L1yeQidadg6wasSwmaK2s8JxFbEXQ0e3AW9KE1xQqmOn0WtOi3lxloStmSeeg==",
    databaseId = "ToDoList",
    collectionId = "Items",
    documentId = "fe7718ad-0000-4f42-cf5a-e2d79d2156df"
  )
print(deleteResult$requestCharge)
}
