% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCollections.R
\name{getCollections}
\alias{getCollections}
\title{Gets a list of collections in a Cosmos DB database}
\usage{
getCollections(connectionInfo, userAgent = "")
}
\arguments{
\item{connectionInfo}{A DocumentDB connection info object generated with getDocumentDBConnectionInfo().}

\item{userAgent}{Optional. A string that specifies the client user agent performing the request. The recommended format is {user agent name}/{version}. For example, the official DocumentDB .NET SDK sets the User-Agent string to Microsoft.Document.Client/1.0.0.0. A custom user-agent could be something like ContosoMarketingApp/1.0.0.}
}
\value{
The information extracted from the REST API response about the collections and some additional details such as request charge and session token.
}
\description{
Gets a list of collections in a Cosmos DB database
}
\examples{
# load the documentdbr package
library(documentdbr)

# get a DocumentDBConnectionInfo object
myConnection <- getDocumentDBConnectionInfo(
  accountUrl = "https://somedocumentdbaccount.documents.azure.com",
  primaryOrSecondaryKey = "t0C36UstTJ4c6vdkFyImkaoB6L1yeQidadg6wasSwmaK2s8JxFbEXQ0e3AW9KE1xQqmOn0WtOi3lxloStmSeeg==",
  databaseId = "MyDatabaseId",
  collectionId = ""  # collectionId is ignored for this operation
)

# get a list of collections and print some info about it
getResult <- getCollections(myConnection)
str(getResult$collections)
print(getResult$collections)

# print request charge
print(getResult$requestCharge)
}
