% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDocument.R
\name{getDocument}
\alias{getDocument}
\title{Gets a single document from a collection.}
\usage{
getDocument(connectionInfo, documentId, partitionKey = "",
  consistencyLevel = "", sessionToken = "", userAgent = "")
}
\arguments{
\item{connectionInfo}{A DocumentDB connection info object generated with getDocumentDbConnectionInfo().}

\item{documentId}{The ID of the document to be queried.}

\item{partitionKey}{Optional. The partition key value for the document to be read. Must be included if and only if the collection is created with a partitionKey definition.}

\item{consistencyLevel}{Optional. The consistency level override. The valid values are: Strong, Bounded, Session, or Eventual (in order of strongest to weakest). The override must be the same or weaker than the account's configured consistency level.}

\item{sessionToken}{Optional. A string token used with session level consistency. For more information, see \href{https://azure.microsoft.com/en-us/documentation/articles/documentdb-consistency-levels}{Using consistency levels in DocumentDB}.}

\item{userAgent}{Optional. A string that specifies the client user agent performing the request. The recommended format is {user agent name}/{version}. For example, the official DocumentDB .NET SDK sets the User-Agent string to Microsoft.Document.Client/1.0.0.0. A custom user-agent could be something like ContosoMarketingApp/1.0.0.}
}
\value{
The result of the query as data.frame object and some information extracted from the REST API response such as request charge and session token.
}
\description{
Gets a single document from a collection.
}
\examples{
# load the documentdbr package
library(documentdbr)

# get a DocumentDbConnectionInfo object
myCollection <- getDocumentDbConnectionInfo(
  accountUrl = "https://somedocumentdbaccount.documents.azure.com",
  primaryOrSecondaryKey = "t0C36UstTJ4c6vdkFyImkaoB6L1yeQidadg6wasSwmaK2s8JxFbEXQ0e3AW9KE1xQqmOn0WtOi3lxloStmSeeg==",
  databaseId = "MyDatabaseId",
  collectionId = "MyCollectionId"
)

# get a single document with a specific id and print some infos about it
getResult <- getDocument(myCollection, documentId = "3c7718ab-858b-4f42-bf5a-e2d79d2156de")
str(getResult$document)
print(getResult$document)

# print request charge
print(getResult$requestCharge)
}
