% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDocumentDBConnectionInfo.R
\name{getDocumentDBConnectionInfo}
\alias{getDocumentDBConnectionInfo}
\title{Returns an object which contains connection information to a DocumentDB collection.}
\usage{
getDocumentDBConnectionInfo(accountUrl, primaryOrSecondaryKey,
  databaseId = "", collectionId = "")
}
\arguments{
\item{accountUrl}{The URI of the DocumentDB account.}

\item{primaryOrSecondaryKey}{The master key to authenticate.}

\item{databaseId}{Optional. The ID of the database to work with. Not needed if operations are on account level.}

\item{collectionId}{Optional. The ID of the collection to work with. Not needed if operations are on database level.}
}
\value{
An object that can be passed to other functions of this package.
}
\description{
Returns an object which contains connection information to a DocumentDB collection.
}
\examples{
# get a DocumentDBConnectionInfo object
myDocDB <- getDocumentDBConnectionInfo(
  accountUrl = "https://somedocumentdbaccount.documents.azure.com",
  primaryOrSecondaryKey = "t0C36UstTJ4c6vdkFyImkaoB6L1yeQidadg6wasSwmaK2s8JxFbEXQ0e3AW9KE1xQqmOn0WtOi3lxloStmSeeg==",
  databaseId = "MyDatabaseId",
  collectionId = "MyCollectionId"
)
}
