% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOffers.R
\name{getOffers}
\alias{getOffers}
\title{Gets a list of available "offers", which are the performance levels you can
provision for a Cosmos DB collection.}
\usage{
getOffers(connectionInfo, userAgent = "")
}
\arguments{
\item{connectionInfo}{A DocumentDB connection info object generated with getDocumentDBConnectionInfo().}

\item{userAgent}{Optional. A string that specifies the client user agent performing the request. The recommended format is {user agent name}/{version}. For example, the official DocumentDB .NET SDK sets the User-Agent string to Microsoft.Document.Client/1.0.0.0. A custom user-agent could be something like ContosoMarketingApp/1.0.0.}
}
\value{
The available offers as a data.frame object and some information extracted from the REST API response such as request charge and session token.
}
\description{
Gets a list of available "offers", which are the performance levels you can
provision for a Cosmos DB collection.
}
\examples{
# load the documentdbr package
library(documentdbr)

# get a DocumentDBConnectionInfo object
myCollection <- getDocumentDBConnectionInfo(
  accountUrl = "https://somedocumentdbaccount.documents.azure.com",
  primaryOrSecondaryKey = "t0C36UstTJ4c6vdkFyImkaoB6L1yeQidadg6wasSwmaK2s8JxFbEXQ0e3AW9KE1xQqmOn0WtOi3lxloStmSeeg==",
  databaseId = "MyDatabaseId",
  collectionId = "MyCollectionId"
)

# get a list of offers and print some info about it
getResult <- getOffers(myCollection)
str(getResult$offers)
print(getResult$offers)

# print request charge
print(getResult$requestCharge)
}
