% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryDocuments.R
\name{queryDocuments}
\alias{queryDocuments}
\title{Queries a collection for documents.}
\usage{
queryDocuments(accountUrl, primaryOrSecondaryKey, databaseId, collectionId,
  queryText, maxItemsPerChunk = 100)
}
\arguments{
\item{accountUrl}{The URI of the DocumentDB account.}

\item{primaryOrSecondaryKey}{The master key to authenticate.}

\item{databaseId}{The ID of the database to query.}

\item{collectionId}{The ID of the collection to query.}

\item{queryText}{The SQL query to execute.}

\item{maxItemsPerChunk}{Optional. Use it for performance and cost tuning.}
}
\value{
The result of the query as data.frame object and some information extracted from the REST API response such as request charge and session token.
}
\description{
Queries a collection for documents.
}
\examples{
# returns the result of a SQL query as data.frame
library(documentdbr)
queryResult <-
  queryDocuments(
    accountUrl = "https://somedocumentdbaccount.documents.azure.com",
    primaryOrSecondaryKey = "t0C36UstTJ4c6vdkFyImkaoB6L1yeQidadg6wasSwmaK2s8JxFbEXQ0e3AW9KE1xQqmOn0WtOi3lxloStmSeeg==",
    databaseId = "ToDoList",
    collectionId = "Items",
    queryText = "SELECT Items.name, Items.description, Items.isComplete FROM Items WHERE Items.isComplete = true",
    maxItemsPerChunk = 100
  )
str(queryResult$documents)
print(queryResult$documents)
print(queryResult$requestCharge)
}
