% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSingleDocument.R
\name{selectSingleDocument}
\alias{selectSingleDocument}
\title{Queries a single document from a collection.}
\usage{
selectSingleDocument(accountUrl, primaryOrSecondaryKey, databaseId,
  collectionId, documentId, partitionKey = "")
}
\arguments{
\item{accountUrl}{The URI of the DocumentDB account.}

\item{primaryOrSecondaryKey}{The master key to authenticate.}

\item{databaseId}{The ID of the database to query.}

\item{collectionId}{The ID of the collection to query.}

\item{documentId}{The ID of the document to be queried.}

\item{partitionKey}{Optional. The partition key value for the document to be read. Must be included if and only if the collection is created with a partitionKey definition.}
}
\value{
The result of the query as data.frame object and some information extracted from the REST API response such as request charge and session token.
}
\description{
Queries a single document from a collection.
}
\examples{
# queries a single document
queryResult <-
  selectSingleDocument(
    accountUrl = "https://somedocumentdbaccount.documents.azure.com",
    primaryOrSecondaryKey = "t0C36UstTJ4c6vdkFyImkaoB6L1yeQidadg6wasSwmaK2s8JxFbEXQ0e3AW9KE1xQqmOn0WtOi3lxloStmSeeg==",
    databaseId = "ToDoList",
    collectionId = "Items",
    documentId = "3c7718ab-858b-4f42-bf5a-e2d79d2156de"
  )
str(queryResult$document)
print(queryResult$document)
print(queryResult$requestCharge)
}
