% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsertDocument.R
\name{upsertDocument}
\alias{upsertDocument}
\title{Inserts or updates the specified JSON document into a collection.}
\usage{
upsertDocument(accountUrl, primaryOrSecondaryKey, databaseId, collectionId,
  document, partitionKey = "")
}
\arguments{
\item{accountUrl}{The URI of the DocumentDB account.}

\item{primaryOrSecondaryKey}{The master key to authenticate.}

\item{databaseId}{The ID of the database to query.}

\item{collectionId}{The ID of the collection to query.}

\item{document}{Either a JSON document (jsonlite) or JSON string denoting the document to add or update.}

\item{partitionKey}{Optional. The partition key value pointing to the partition where the document is (to be) stored.}
}
\value{
Some information extracted from the REST API response such as request charge and session token.
}
\description{
Inserts or updates the specified JSON document into a collection.
}
\examples{
# inserts a document if not existing yet, otherwise updates it
library(documentdbr)
upsertResult <-
  upsertDocument(
    accountUrl = "https://somedocumentdbaccount.documents.azure.com",
    primaryOrSecondaryKey = "t0C36UstTJ4c6vdkFyImkaoB6L1yeQidadg6wasSwmaK2s8JxFbEXQ0e3AW9KE1xQqmOn0WtOi3lxloStmSeeg==",
    databaseId = "ToDoList",
    collectionId = "Items",
    document = toJSON("{\\"id\\": \\"fe7718ad-0000-4f42-cf5a-e2d79d2156df\\", \\"Value\\": \\"126\\"}")
  )
print(upsertResult$requestCharge)
}
