% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteDocument.R
\name{deleteDocument}
\alias{deleteDocument}
\title{Deletes a single document from a collection.}
\usage{
deleteDocument(connectionInfo, documentId, partitionKey = "",
  consistencyLevel = "", sessionToken = "", userAgent = "")
}
\arguments{
\item{connectionInfo}{A DocumentDB connection info object generated with getDocumentDBConnectionInfo().}

\item{documentId}{The ID of the document to delete.}

\item{partitionKey}{Optional. The partition key value for the document to be deleted. Must be included if and only if the collection is created with a partitionKey definition.}

\item{consistencyLevel}{Optional. The consistency level override. The valid values are: Strong, Bounded, Session, or Eventual (in order of strongest to weakest). The override must be the same or weaker than the account's configured consistency level.}

\item{sessionToken}{Optional. A string token used with session level consistency. For more information, see \href{https://azure.microsoft.com/en-us/documentation/articles/documentdb-consistency-levels}{Using consistency levels in DocumentDB}.}

\item{userAgent}{Optional. A string that specifies the client user agent performing the request. The recommended format is {user agent name}/{version}. For example, the official DocumentDB .NET SDK sets the User-Agent string to Microsoft.Document.Client/1.0.0.0. A custom user-agent could be something like ContosoMarketingApp/1.0.0.}
}
\value{
Some information extracted from the REST API response such as request charge and session token.
}
\description{
Deletes a single document from a collection.
}
\examples{
# load the documentdbr package
library(documentdbr)

# get a DocumentDBConnectionInfo object
myCollection <- getDocumentDBConnectionInfo(
  accountUrl = "https://somedocumentdbaccount.documents.azure.com",
  primaryOrSecondaryKey = "t0C36UstTJ4c6vdkFyImkaoB6L1yeQidadg6wasSwmaK2s8JxFbEXQ0e3AW9KE1xQqmOn0WtOi3lxloStmSeeg==",
  databaseId = "MyDatabaseId",
  collectionId = "MyCollectionId"
)

# delete the document with id fe7718ad-0000-4f42-cf5a-e2d79d2156df
deleteResult <- deleteDocument(myCollection, documentId = "fe7718ad-0000-4f42-cf5a-e2d79d2156df")

# print the request charge
print(deleteResult$requestCharge)
}
