% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDocumentDbConnectionInfo.R
\name{getDocumentDbConnectionInfo}
\alias{getDocumentDbConnectionInfo}
\title{Returns an object which contains connection information to a DocumentDB collection.}
\usage{
getDocumentDbConnectionInfo(accountUrl, primaryOrSecondaryKey, databaseId,
  collectionId)
}
\arguments{
\item{accountUrl}{The URI of the DocumentDB account.}

\item{primaryOrSecondaryKey}{The master key to authenticate.}

\item{databaseId}{The ID of the database to query/modify.}

\item{collectionId}{The ID of the collection to query/modify.}
}
\value{
An object that can be passed to other functions of this package.
}
\description{
Returns an object which contains connection information to a DocumentDB collection.
}
\examples{
# get a DocumentDbConnectionInfo object
myDocDB <- getDocumentDbConnectionInfo(
  accountUrl = "https://somedocumentdbaccount.documents.azure.com",
  primaryOrSecondaryKey = "t0C36UstTJ4c6vdkFyImkaoB6L1yeQidadg6wasSwmaK2s8JxFbEXQ0e3AW9KE1xQqmOn0WtOi3lxloStmSeeg==",
  databaseId = "MyDatabaseId",
  collectionId = "MyCollectionId"
)
}
