% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSingleDocument.R
\name{selectSingleDocument}
\alias{selectSingleDocument}
\title{Queries a single document from a collection.}
\usage{
selectSingleDocument(accountUrl, primaryOrSecondaryKey, databaseId,
  collectionId, documentId, partitionKey = "", consistencyLevel = "",
  sessionToken = "", userAgent = "")
}
\arguments{
\item{accountUrl}{The URI of the DocumentDB account.}

\item{primaryOrSecondaryKey}{The master key to authenticate.}

\item{databaseId}{The ID of the database to query.}

\item{collectionId}{The ID of the collection to query.}

\item{documentId}{The ID of the document to be queried.}

\item{partitionKey}{Optional. The partition key value for the document to be read. Must be included if and only if the collection is created with a partitionKey definition.}

\item{consistencyLevel}{Optional. The consistency level override. The valid values are: Strong, Bounded, Session, or Eventual (in order of strongest to weakest). The override must be the same or weaker than the account's configured consistency level.}

\item{sessionToken}{Optional. A string token used with session level consistency. For more information, see \href{https://azure.microsoft.com/en-us/documentation/articles/documentdb-consistency-levels}{Using consistency levels in DocumentDB}.}

\item{userAgent}{Optional. A string that specifies the client user agent performing the request. The recommended format is {user agent name}/{version}. For example, the official DocumentDB .NET SDK sets the User-Agent string to Microsoft.Document.Client/1.0.0.0. A custom user-agent could be something like ContosoMarketingApp/1.0.0.}
}
\value{
The result of the query as data.frame object and some information extracted from the REST API response such as request charge and session token.
}
\description{
Queries a single document from a collection.
}
\examples{
# queries a single document
queryResult <-
  selectSingleDocument(
    accountUrl = "https://somedocumentdbaccount.documents.azure.com",
    primaryOrSecondaryKey = "t0C36UstTJ4c6vdkFyImkaoB6L1yeQidadg6wasSwmaK2s8JxFbEXQ0e3AW9KE1xQqmOn0WtOi3lxloStmSeeg==",
    databaseId = "ToDoList",
    collectionId = "Items",
    documentId = "3c7718ab-858b-4f42-bf5a-e2d79d2156de"
  )
str(queryResult$document)
print(queryResult$document)
print(queryResult$requestCharge)
}
